Safari-611.1.10/LayoutTests/printing/page-break-margin-collapsed.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/paged-media-test-utils.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="sandbox"></div>
<script>
description("Test for big margin.");

function test()
{
    createBlockWithRatioToPageHeight("page1", 0.1).style.pageBreakBefore = "always";
    createBlockWithRatioToPageHeight("page2", 0.1).style.marginTop = "100000px";

    pageNumberForElementShouldBe("page1", 1);
    // Instead of inserting many empty pages, we should place block 'page2'
    // on the next page, collapsing the margin between vetically adjacent blocks.
    pageNumberForElementShouldBe("page2", 2);

    document.body.removeChild(document.getElementById("sandbox"));
}

var successfullyParsed = true;
</script>
<script>runPrintingTest(test);</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
