LayoutTests/printing/page-break-margin-collapsed-expected.txt

Test for big margin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS: page number of "page1" is 1
FAIL: expected page number of "page2" is 2. Was 167

PASS successfullyParsed is true

TEST COMPLETE
